+++ src/Streamly/Internal/Prelude.hs
@@ -447,7 +447,7 @@ import Streamly.Internal.Data.Unfold.Types (Unfold)
 import Streamly.Internal.Memory.Array.Types (Array, writeNUnsafe)
 -- import Streamly.Memory.Ring (Ring)
 import Streamly.Internal.Data.SVar (MonadAsync, State, defState, adaptState)
-import Streamly.Streams.Async (mkAsync')
+import Streamly.Internal.Data.Stream.Async (mkAsync')
 import Streamly.Streams.Combinators (inspectMode, maxYields)
 import Streamly.Streams.Prelude
        (fromStreamS, toStreamS, foldWith, foldMapWith, forEachWith)
